♻️ JE-115-refactor: 주소, 설명 컴포넌트 분리
