+++ src/styles/globals.css
@@ -186,7 +186,3 @@ input:-webkit-autofill:active {
   margin-bottom: 8px;
   color: #112211;
 }
-
-.custom-editor {
-  height: 350px !important;
-}
